requirements.txt
streamlit
scikit-learn
opencv-python
numpy
joblib
Pillow